DMSApp/serializers.py
from rest_framework import serializers

from .models import Document, Entity


class EntitySerializer(serializers.ModelSerializer):
    class Meta:
        model = Entity
        fields = "__all__"


class DocumentSerializer(serializers.ModelSerializer):
    entities = EntitySerializer(many=True, read_only=True)

    class Meta:
        model = Document
        fields = "__all__"
