requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
click==8.1.7
colorama==0.4.6
Django==4.2.16
django-cors-headers==4.6.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
drf-yasg==1.21.8
gunicorn==23.0.0
h11==0.14.0
inflection==0.5.1
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pytest==8.3.3
pytest-django==4.9.0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
soupsieve==2.6
sqlparse==0.5.2
tzdata==2024.2
uritemplate==4.1.1
uvicorn==0.32.1
